+++ examples/topology/3dBorderExtraction.cpp
@@ -45,8 +45,7 @@
 #include "DGtal/base/Common.h"
 
 #include "DGtal/io/readers/VolReader.h"
-#include "DGtal/io/DrawWithDisplay3DModifier.h"
-#include "DGtal/io/viewers/Viewer3D.h"
+#include "DGtal/io/viewers/PolyscopeViewer.h"
 #include "DGtal/io/Color.h"
 #include "DGtal/images/ImageSelector.h"
 #include "DGtal/helpers/StdDefs.h"
@@ -95,15 +94,14 @@ int main( int argc, char** argv )
  ObjectType bdiamond = diamond.border(); // one component
  ObjectType bdiamond_clone = diamond_clone.border(); // two components
 
- QApplication application(argc,argv);
- typedef Viewer3D<> MyViewer;
+ typedef PolyscopeViewer<> MyViewer;
  MyViewer viewer;
- viewer.show();
- viewer<<  CustomColors3D(Color(250, 250,250),Color(250, 250,250));
+ viewer<<  Color(250, 250,250);
  viewer << bdiamond_clone;
  viewer << bdiamond ;
- viewer << ClippingPlane(1,1,0,5, false) << MyViewer::updateDisplay;
- return application.exec();
+ viewer << ClippingPlane(1,1,0,5);
+ viewer.show();
+ return 0;
 
 }
 //                                                                           //