+++ tests/io/viewers/test3DImage3DView.cpp
@@ -22,7 +22,7 @@
  *
  * @date 2013/04/29
  *
- * Functions for testing class Viewer3D.
+ * Functions for testing class PolyscopeViewer.
  *
  * This file is part of the DGtal library.
  */
@@ -30,15 +30,14 @@
 ///////////////////////////////////////////////////////////////////////////////
 #include <iostream>
 #include "DGtal/base/Common.h"
-#include "DGtal/io/viewers/Viewer3D.h"
-#include "DGtal/io/DrawWithDisplay3DModifier.h"
+#include "DGtal/io/viewers/PolyscopeViewer.h"
 #include "DGtal/io/Color.h"
 #include "DGtal/io/readers/VolReader.h"
 #include "DGtal/helpers/StdDefs.h"
 #include "DGtal/shapes/Shapes.h"
 #include "DGtal/math/BasicMathFunctions.h"
+#include "DGtal/kernel/BasicPointFunctors.h"
 #include "ConfigTest.h"
-#include "DGtal/io/colormaps/HueShadeColorMap.h"
 
 
 ///////////////////////////////////////////////////////////////////////////////
@@ -48,19 +47,6 @@ using namespace DGtal;
 using namespace Z3i;
 
 
-
-///////////////////////////////////////////////////////////////////////////////
-// Functions for testing class Viewer3D.
-///////////////////////////////////////////////////////////////////////////////
-struct hueFct{
-inline
- unsigned int operator() (unsigned int aVal) const
-  {
-    HueShadeColorMap<unsigned int>  hueShade(0,255);
-    DGtal::Color col = hueShade(aVal);
-    return  col.getRGB();
-  }
-};
 ///////////////////////////////////////////////////////////////////////////////
 // Standard services - public :
 
@@ -68,49 +54,41 @@ int main( int argc, char** argv )
 {
   typedef DGtal::ImageContainerBySTLVector< DGtal::Z3i::Domain, unsigned char>  Image3D;
 
- QApplication application(argc,argv);
- Viewer3D<> viewer;
- viewer.setWindowTitle("simpleViewer");
- viewer.show();
- trace.beginBlock("Testing Viewer with display of 3D Image  ");
+  PolyscopeViewer viewer;
+  trace.beginBlock("Testing Viewer with display of 3D Image  ");
 
- Point p1( 0, 0, 0 );
- Point p2( 125, 188, 0 );
- Point p3( 30, 30, 30 );
+  Point p1( 0, 0, 0 );
+  Point p2( 125, 188, 0 );
+  Point p3( 30, 30, 30 );
 
- std::string filename =  testPath + "samples/lobsterCroped.vol";
- hueFct huefct;
+  std::string filename =  testPath + "samples/lobsterCroped.vol";
 
- viewer.setFillTransparency(150);
- Image3D image3d =  VolReader<Image3D>::importVol(filename);
- viewer << SetMode3D(image3d.className(), "BoundingBox");
+  Image3D image3d =  VolReader<Image3D>::importVol(filename);
+  
+  viewer << Color(255, 255, 200);
+  viewer << image3d;
+  // Extract some slice images:
+  // Get the 2D domain of the slice:
+  DGtal::functors::Projector<DGtal::Z2i::Space>  invFunctor; invFunctor.initRemoveOneDim(2);
+  DGtal::Z2i::Domain domain2D(invFunctor(image3d.domain().lowerBound()),
+         invFunctor(image3d.domain().upperBound()));
 
- viewer << DGtal::AddTextureImage3DWithFunctor<Image3D,  hueFct , Space, KSpace>(image3d, huefct, Viewer3D<>::RGBMode );
- viewer.setFillTransparency(255);
- // Extract some slice images:
- // Get the 2D domain of the slice:
- DGtal::functors::Projector<DGtal::Z2i::Space>  invFunctor; invFunctor.initRemoveOneDim(2);
- DGtal::Z2i::Domain domain2D(invFunctor(image3d.domain().lowerBound()),
-           invFunctor(image3d.domain().upperBound()));
+  typedef DGtal::ConstImageAdapter<Image3D, DGtal::Z2i::Domain,  DGtal::functors::Projector< Z3i::Space>,
+                                Image3D::Value,  functors::Identity >  SliceImageAdapter;
+  functors::Identity idV;
+  functors::Projector<DGtal::Z3i::Space> aSliceFunctorZ(5); aSliceFunctorZ.initAddOneDim(2);
 
- typedef DGtal::ConstImageAdapter<Image3D, DGtal::Z2i::Domain,  DGtal::functors::Projector< Z3i::Space>,
-                                  Image3D::Value,  functors::Identity >  SliceImageAdapter;
- functors::Identity idV;
- functors::Projector<DGtal::Z3i::Space> aSliceFunctorZ(5); aSliceFunctorZ.initAddOneDim(2);
+  SliceImageAdapter sliceImageZ(image3d, domain2D, aSliceFunctorZ, idV);
 
- SliceImageAdapter sliceImageZ(image3d, domain2D, aSliceFunctorZ, idV);
+  std::string name = viewer.draw(sliceImageZ);
+  viewer.data[name].transform.translate(Eigen::Vector3d(0, 0, -10));
 
-  viewer << sliceImageZ;
-  viewer <<  DGtal::UpdateImagePosition<Space, KSpace>(6, Viewer3D<>::zDirection, 0.0, 0.0, -10.0);
+  viewer << p1 << p2 << p3;
 
- viewer << p1 << p2 << p3;
- viewer << Viewer3D<>::updateDisplay;
-
-
- bool res = application.exec();
- trace.emphase() << ( res ? "Passed." : "Error." ) << endl;
- trace.endBlock();
- return res ? 0 : 1;
+  trace.emphase() << "Passed." << endl;
+  trace.endBlock();
+  viewer.show();
+  return 0;
 
 
 }