+++ ChangeLog.md
@@ -7,6 +7,8 @@
   - Fix compilation warnings (Bastien Doignies, [#1757](https://github.com/DGtal-team/DGtal/pull/))
   - Remove GMP dependency (Bastien Doignies, [#1769](https://github.com/DGtal-team/DGtal/pull/1769))
   - New Boost BigInteger Backend (Bastien Doignies, [#1769](https://github.com/DGtal-team/DGtal/pull/1769))
+  - New Polyscope-based polyscope (Bastien Doignies, [#1775](https://github.com/DGtal-team/DGtal/pull/1775))
+  - Remove old QGLViewer and Qt dependancy (Bastien Doignies, [#1775](https://github.com/DGtal-team/DGtal/pull/1775))
   - Remove deprecated classes and functions (Bastien DOIGNIES, [#1765](https://github.com/DGtal-team/DGtal/pull/1765))
   - Boost is no longer required when building DGtal, boost is now fetched using CPM at cmake step) (Bastien Doignies, David Coeurjolly,  [#1763](https://github.com/DGtal-team/DGtal/pull/1763))
   - Patate upgraded to ponca (Bastien Doignies, [1776](https://github.com/DGtal-team/DGtal/pull/1776))